Move destinations into a separate file
