+++ package.json
@@ -19,7 +19,8 @@
   },
   "homepage": "https://github.com/majac91/Travel_journal#readme",
   "dependencies": {
-    "moment": "^2.27.0"
+    "moment": "^2.27.0",
+    "src": "^1.1.2"
   },
   "devDependencies": {
     "sass": "^1.26.10"
